scripts/telethon/requirements.txt
telethon==1.41.1
python-dotenv==1.1.1
psycopg2-binary==2.9.9
requests==2.32.3
aiofiles==23.2.1